RDCC-178 package refactor for persistense and fixed check style validations
