+++ src/integrationTest/java/uk/gov/hmcts/reform/professionalapi/CreateOrganisationWithPaymentAccountTest.java
@@ -20,10 +20,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import uk.gov.hmcts.reform.professionalapi.controller.request.OrganisationCreationRequest;
 import uk.gov.hmcts.reform.professionalapi.domain.Organisation;
 import uk.gov.hmcts.reform.professionalapi.domain.PaymentAccount;
-import uk.gov.hmcts.reform.professionalapi.service.persistence.ContactInformationRepository;
-import uk.gov.hmcts.reform.professionalapi.service.persistence.OrganisationRepository;
-import uk.gov.hmcts.reform.professionalapi.service.persistence.PaymentAccountRepository;
-import uk.gov.hmcts.reform.professionalapi.service.persistence.ProfessionalUserRepository;
+import uk.gov.hmcts.reform.professionalapi.persistence.ContactInformationRepository;
+import uk.gov.hmcts.reform.professionalapi.persistence.OrganisationRepository;
+import uk.gov.hmcts.reform.professionalapi.persistence.PaymentAccountRepository;
+import uk.gov.hmcts.reform.professionalapi.persistence.ProfessionalUserRepository;
 import uk.gov.hmcts.reform.professionalapi.util.ProfessionalReferenceDataClient;
 import uk.gov.hmcts.reform.professionalapi.util.Service2ServiceEnabledIntegrationTest;
